Changed property Sphere.R from double to float to be more consistent with the rest of the code

+++ Raytracing/Sphere.cs
@@ -3,10 +3,10 @@
 namespace Raytracing {
     class Sphere : ISceneObject {
         public Vector3 Centre { get; }
-        public double R { get; }
+        public float R { get; }
         public Colour Colour { get; }
 
-        public Sphere(Vector3 centre, double r, Colour colour) {
+        public Sphere(Vector3 centre, float r, Colour colour) {
             this.Centre = centre;
             this.R = r;
             this.Colour = colour;
